rag_agent/requirements.txt
langchain
langchain-core
langchain-community
langchain-text-splitters
langchain-huggingface

faiss-cpu
sentence-transformers
rank-bm25
gitpython

google-generativeai
protobuf==4.25.3
requests
typing_extensions

python-dotenv
numpy